Request types

+++ src/main/java/com/keyin/flight/FlightRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FlightRepository extends CrudRepository<Flight, Long> {
+public interface FlightRepository extends CrudRepository<Flight, Integer> {
 }
